+++ lib/src/state_stream/index.dart
@@ -1,2 +1,3 @@
 export 'state_stream.dart';
 export 'state_stream_impl.dart';
+export 'state_subject.dart';